Moved field/beam types to MagFieldParam and extracted it to separate module
